Fixed 64 bit issue in field Handle definition of DIR structure.

+++ trunk/etw/include/highdirent.h
@@ -27,7 +27,7 @@ struct dirent                /* data from readdir() */
 
 
 typedef struct {
-    long Handle;
+    intptr_t Handle;
     char DirName[257];
     struct dirent DirentBuf;
     struct _finddata_t FindBuf;
